+++ test/Main.hs
@@ -15,6 +15,8 @@ import Plutarch (
  )
 import Plutarch.ByteString (pallBS)
 import Plutarch.Evaluate (evalTerm)
+import Plutarch.Lift (PLifted, PUnsafeLiftDecl)
+import Plutarch.List (pcheckSorted, preverse)
 import Plutarch.Prelude (
   PBool,
   PUnit,
@@ -69,6 +71,29 @@ main = do
                 (pallBS # pisHexDigit # pconstant "abcdef")
             , testCase "no other ASCII code is a hex digit" . isTrue $
                 (pallBS # plam (\x -> pnot #$ pisHexDigit # x) # pconstant nonHexAscii)
+            ]
+        ]
+    , testGroup
+        "List utilities"
+        [ testGroup
+            "preverse"
+            [ testCase "reversing an empty list is identity" . matches (pconstant ([] :: [Integer])) $
+                (preverse # pconstant ([] :: [Integer]))
+            , testCase "reversing a singleton is identity" . matches (pconstant ([1] :: [Integer])) $
+                (preverse # pconstant ([1] :: [Integer]))
+            , testCase "reversing a non-singleton" . matches (pconstant ([2, 1] :: [Integer])) $
+                (preverse # pconstant ([1, 2] :: [Integer]))
+            ]
+        , testGroup
+            "pcheckSorted"
+            [ testCase "empty lists are sorted" . isTrue $
+                (pcheckSorted # pconstant ([] :: [Integer]))
+            , testCase "singleton lists are sorted" . isTrue $
+                (pcheckSorted # pconstant ([1] :: [Integer]))
+            , testCase "two items in the right order are sorted" . isTrue $
+                (pcheckSorted # pconstant ([1, 2] :: [Integer]))
+            , testCase "two items in the wrong order are not sorted" . isFalse $
+                (pcheckSorted # pconstant ([2, 1] :: [Integer]))
             ]
         ]
     , testGroup
@@ -189,6 +214,24 @@ isFalse comp = case evalTerm NoTracing comp of
   Right (res, _, _) -> case res of
     Left err -> assertFailure $ "Execution errored: " <> show err
     Right t -> when (plift t) (assertFailure "is true")
+
+matches ::
+  forall (a :: S -> Type).
+  (PUnsafeLiftDecl a, Eq (PLifted a)) =>
+  -- expected
+  (forall (s :: S). Term s a) ->
+  -- actual
+  (forall (s :: S). Term s a) ->
+  IO ()
+matches expected actual = case evalTerm NoTracing expected of
+  Left err -> assertFailure $ "Expected case failed to compile: " <> Text.unpack err
+  Right (res, _, _) -> case res of
+    Left err -> assertFailure $ "Expected case execution errored: " <> show err
+    Right expected' -> case evalTerm NoTracing actual of
+      Left err -> assertFailure $ "Case did not compile: " <> Text.unpack err
+      Right (res', _, _) -> case res' of
+        Left err -> assertFailure $ "Case execution errored: " <> show err
+        Right actual' -> unless (plift expected' == plift actual') (assertFailure "cases do not match")
 
 nonHexAscii :: ByteString
 nonHexAscii =